requirements.txt
torch
transformers
soundfile
sounddevice
numpy
flask
openai-whisper
python-dotenv
openai
dotenv
requests